+++ changelog.md
@@ -2,6 +2,8 @@
 
 A User's UserProfile is now automatically created when you create a user in a post save signal.
 
+The date of birth template tag now allows you have pass in a label to be displayed next to the form.
+
 ### 0.13.1 (Minor Release)
 
 Upgrades the setup.py Django version from 2.0.9 to 2.0.13. Removes the six library dependency from setup.py.