+++ changelog.md
@@ -2,6 +2,8 @@
 
 A User's UserProfile is now automatically created when you create a user in a post save signal.
 
+Changes Diagnosis.details, Allergies.details. and PastMedicalHistory.details from char fields to a text fields. Migrations will have to be run, existing data will be maintained.
+
 ### 0.13.1 (Minor Release)
 
 Upgrades the setup.py Django version from 2.0.9 to 2.0.13. Removes the six library dependency from setup.py.