+++ changelog.md
@@ -5,6 +5,11 @@ A User's UserProfile is now automatically created when you create a user in a po
 RecordEditor.editItem, RecordEditor.editItem and RecordEditor.openEditItemModal to take a url argument.
 This is the url of the template that should be opened with the edit item controller.
 
+RecordEditor.editItem also will now take an item or an index rather than just an index.
+Taking an index has been deprecated and will be removed in v0.15.0
+
+The PatientListCtrl now does the logic of whether to create a new item or edit an item within the
+controller before it calls the record editor.
 
 ### 0.13.1 (Minor Release)